utf8

+++ varints/base128.h
@@ -9,10 +9,10 @@ int b128_unpack( const struct mm_block *container, size_t *result );
 int b128_unpack2( const void *data, size_t lenght, size_t *result );
 int b128_unpack_shift( const void **data, size_t *lenght, size_t *result );
 
-int b128_pack( size_t number, struct mm_block *container );
-int b128_pack2( size_t number, void *container, size_t avail );
-int b128_pack_shift( size_t number, void **container, size_t *avail );
+size_t b128_pack( size_t number, struct mm_block *container );
+size_t b128_pack2( size_t number, void *container, size_t avail );
+size_t b128_pack_shift( size_t number, void **container, size_t *avail );
 
-int b128_pack_append( size_t number, struct mm_block *container );
+size_t b128_pack_append( size_t number, struct mm_block *container );
 
 #endif // BASE128_H
